Use new session persistence scheme
Instead of randomly generated keys to an in-memory store of computed
database keys, the sessions are now encrypted versions of the computed
database key. The key used to encrypt the sessions is short-lived, so
risk of compromise should be low. Even if the server keys are
compromised, the database can only be compromised if the attacker has
access to client session cookies as well, but it still never exposes the
actual database passphrase.

Fixes #31